@@ -7,7 +7,7 @@ module.exports = function(api) {
         [
             '@babel/preset-env',
             {
-                targets: api.env('test') ? { node: 'current' } : { chrome: '86' },
+                targets: api.env('test') ? { node: 'current' } : { chrome: '91' },
 
                 //targets: { chrome: '86', node: 'current' },
 
@@ -23,10 +23,6 @@ module.exports = function(api) {
         '@babel/plugin-proposal-export-namespace-from',
         '@babel/plugin-proposal-class-properties',
         ['@babel/plugin-proposal-pipeline-operator', { proposal: 'fsharp' }],
-        //"@babel/plugin-proposal-partial-application",
-        //'@babel/plugin-proposal-nullish-coalescing-operator',
-        //'@babel/plugin-proposal-optional-chaining',
-        //'@babel/plugin-proposal-do-expressions',(()=>{})();
     ]
 
     return { presets, plugins }
