@@ -2,13 +2,12 @@
 import { map } from 'rxjs/operators'
 import { Deep, replaceValueDeep } from 'structural-comparison'
 
-///同combineLatest操作符，不同於合并數組，本函數合并的為Deep對象。
-
 /**
- *
+ * 返回一个Observable实例，组合成同形状的Deep发射，词条值为对应源Observable的值。
  * @param {Deep} deep
  * @returns {Observable}
  */
-export function deepCombineLatest(deep) {
-    return combineLatest(deep.getValues()) |> map(values => deep |> replaceValueDeep(values))
+export function deepCombineLatest(deep = new Deep([])) {
+    return combineLatest(deep.getValues())
+        |> map(values => deep |> replaceValueDeep(values))
 }
